+++ lib/boruta/config.ex
@@ -102,7 +102,22 @@ defmodule Boruta.Config do
   @spec resource_owners() :: module()
   @doc false
   def resource_owners do
-    Keyword.fetch!(oauth_config(), :contexts)[:resource_owners]
+    case Keyword.fetch!(oauth_config(), :contexts)[:resource_owners] do
+      nil ->
+        raise """
+        Missing configuration for resource_owners context. You can set your own
+        `Boruta.Oauth.ResourceOwners` behaviour implementation in config:
+
+          config :boruta, Boruta.Oauth,
+            repo: MyApp.Repo,
+            contexts: [
+              resource_owners: MyApp.ResourceOwners
+            ]
+        """
+
+      module ->
+        module
+    end
   end
 
   @spec issuer() :: String.t()