+++ lib/boruta/oauth.ex
@@ -1,9 +1,24 @@
+defmodule Boruta.OauthModule do
+  @moduledoc false
+
+  alias Boruta.Oauth.ResourceOwner
+
+  @callback token(conn :: Plug.Conn.t() | map(), module :: atom()) :: any()
+  @callback preauthorize(conn :: Plug.Conn.t() | map(), resource_owner :: ResourceOwner.t(), module :: atom()) :: any()
+  @callback authorize(conn :: Plug.Conn.t() | map(), resource_owner :: ResourceOwner.t(), module :: atom()) :: any()
+  @callback introspect(conn :: Plug.Conn.t() | map(), module :: atom()) :: any()
+  @callback revoke(conn :: Plug.Conn.t() | map(), module :: atom()) :: any()
+end
+
 defmodule Boruta.Oauth do
   @moduledoc """
   Boruta OAuth entrypoint, handles OAuth requests.
 
-  Note : this module works in association with `Boruta.Oauth.Application` behaviour
+  > Note : this module follows inverted heaxagonal architecture, its functions will invoke functions of the given argument module as result.
+  > The definition of the callbacks are provided by either `Boruta.Oauth.Application` or `Boruta.Oauth.AuthorizeApplication`, `Boruta.Oauth.TokenApplication`, `Boruta.Oauth.IntrospectApplication`, and `Boruta.Oauth.RevokeApplication`,
   """
+
+  @behaviour Boruta.OauthModule
 
   alias Boruta.Oauth.Authorization
   alias Boruta.Oauth.AuthorizeResponse
@@ -21,6 +36,7 @@ defmodule Boruta.Oauth do
   Triggers `token_success` in case of success and `token_error` in case of failure from the given `module`. Those functions are described in `Boruta.Oauth.Application` behaviour.
   """
   @spec token(conn :: Plug.Conn.t() | map(), module :: atom()) :: any()
+  @impl true
   def token(%Plug.Conn{} = conn, module) when is_atom(module) do
     with {:ok, request} <- Request.token_request(conn),
          {:ok, tokens} <- Authorization.token(request),
@@ -47,13 +63,9 @@ defmodule Boruta.Oauth do
 
   Triggers `preauthorize_success` in case of success and `preauthorize_error` in case of failure from the given `module`. Those functions are described in `Boruta.Oauth.Application` behaviour.
   """
-  @spec preauthorize(
-          conn :: Plug.Conn.t() | map(),
-          resource_owner :: ResourceOwner.t(),
-          module :: atom()
-        ) :: any()
-  def preauthorize(%Plug.Conn{} = conn, %ResourceOwner{} = resource_owner, module)
-      when is_atom(module) do
+  @spec preauthorize(conn :: Plug.Conn.t() | map(), resource_owner :: ResourceOwner.t(), module :: atom()) :: any()
+  @impl true
+  def preauthorize(%Plug.Conn{} = conn, %ResourceOwner{} = resource_owner, module) when is_atom(module) do
     with {:ok, request} <- Request.authorize_request(conn, resource_owner),
          {:ok, authorization} <- Authorization.preauthorize(request) do
       module.preauthorize_success(
@@ -77,13 +89,9 @@ defmodule Boruta.Oauth do
 
   Triggers `authorize_success` in case of success and `authorize_error` in case of failure from the given `module`. Those functions are described in `Boruta.Oauth.Application` behaviour.
   """
-  @spec authorize(
-          conn :: Plug.Conn.t() | map(),
-          resource_owner :: ResourceOwner.t(),
-          module :: atom()
-        ) :: any()
-  def authorize(%Plug.Conn{} = conn, %ResourceOwner{} = resource_owner, module)
-      when is_atom(module) do
+  @spec authorize(conn :: Plug.Conn.t() | map(), resource_owner :: ResourceOwner.t(), module :: atom()) :: any()
+  @impl true
+  def authorize(%Plug.Conn{} = conn, %ResourceOwner{} = resource_owner, module) when is_atom(module) do
     with {:ok, request} <- Request.authorize_request(conn, resource_owner),
          {:ok, tokens} <- Authorization.token(request),
          %AuthorizeResponse{} = response <- AuthorizeResponse.from_tokens(tokens) do
@@ -122,6 +130,7 @@ defmodule Boruta.Oauth do
   Triggers `introspect_success` in case of success and `introspect_error` in case of failure from the given `module`. Those functions are described in `Boruta.Oauth.Application` behaviour.
   """
   @spec introspect(conn :: Plug.Conn.t() | map(), module :: atom()) :: any()
+  @impl true
   def introspect(%Plug.Conn{} = conn, module) when is_atom(module) do
     with {:ok, request} <- Request.introspect_request(conn),
          {:ok, token} <- Introspect.token(request) do
@@ -141,6 +150,7 @@ defmodule Boruta.Oauth do
   Triggers `revoke_success` in case of success and `revoke_error` in case of failure from the given `module`. Those functions are described in `Boruta.Oauth.Application` behaviour.
   """
   @spec revoke(conn :: Plug.Conn.t() | map(), module :: atom()) :: any()
+  @impl true
   def revoke(%Plug.Conn{} = conn, module) when is_atom(module) do
     with {:ok, request} <- Request.revoke_request(conn),
          :ok <- Revoke.token(request) do