+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - `AuthorizeApplication`, `IntrospectApplication`, `RevokeApplication`, and `TokenApplication` behaviours allowing to implement separatly different OAuth use cases.
     - `list_active_tokens` Ecto admin function
     - `Boruta.AccessTokensAdapter`, `Boruta.CodesAdapter`, `Boruta.ClientsAdapter`, and `Boruta.ScopesAdapter` encapsulating adapters that are set in configuration.
+    - `Boruta.Oauth.AuthorizeResponse.redirect_to_url/1` function
   ### Security
     - do not issue access_tokens from other clients refresh tokens
 ## [1.0.3] - 2021-07-29